Move social feature functions to social_features.py
