+++ discovery-provider/src/tasks/entity_manager/social_features.py
@@ -3,7 +3,15 @@ from typing import Union
 from src.models.social.follow import Follow
 from src.models.social.repost import Repost
 from src.models.social.save import Save
-from src.tasks.entity_manager.utils import Action, EntityType, ManageEntityParameters
+from src.premium_content.premium_content_access_checker import (
+    premium_content_access_checker,
+)
+from src.tasks.entity_manager.utils import (
+    Action,
+    EntityType,
+    ExistingRecordDict,
+    ManageEntityParameters,
+)
 
 action_to_record_type = {
     Action.FOLLOW: EntityType.FOLLOW,
@@ -126,6 +134,64 @@ def delete_social_record(params):
             params.entity_id,
             record_type,
             deleted_record,
+        )
+
+
+def create_social_record_if_access(
+    existing_records: ExistingRecordDict, params: ManageEntityParameters
+):
+    user_id = params.user_id
+    entity_id = params.entity_id
+    entity_type = params.entity_type
+    entity = (
+        existing_records[entity_type][entity_id]
+        if entity_type in existing_records
+        and entity_id in existing_records[entity_type]
+        else None
+    )
+    if entity:
+        premium_content_access = premium_content_access_checker.check_access(
+            {
+                "user_id": user_id,
+                "premium_content_id": entity_id,
+                "premium_content_type": "track",
+                "premium_content": entity,
+            }
+        )
+        if premium_content_access["does_user_have_access"]:
+            create_social_record(params)
+    else:
+        raise Exception(
+            "entity_manager.py | create_social_record_if_access | entity is not in existing records"
+        )
+
+
+def delete_social_record_if_access(
+    existing_records: ExistingRecordDict, params: ManageEntityParameters
+):
+    user_id = params.user_id
+    entity_id = params.entity_id
+    entity_type = params.entity_type
+    entity = (
+        existing_records[entity_type][entity_id]
+        if entity_type in existing_records
+        and entity_id in existing_records[entity_type]
+        else None
+    )
+    if entity:
+        premium_content_access = premium_content_access_checker.check_access(
+            {
+                "user_id": user_id,
+                "premium_content_id": entity_id,
+                "premium_content_type": "track",
+                "premium_content": entity,
+            }
+        )
+        if premium_content_access["does_user_have_access"]:
+            delete_social_record(params)
+    else:
+        raise Exception(
+            "entity_manager.py | delete_social_record_if_access | entity is not in existing records"
         )
